requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
groq

pydantic
python-dotenv
streamlit
ipykernel
requests

fastapi
uvicorn
langserve
#sse-starlette  check if this is fixed or not